Move command and handler to new home
