+++ source/Energinet.DataHub.MeteringPoints.Application/Create/ExchangeDetails.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MeteringPoints.Application.Create.Consumption
+namespace Energinet.DataHub.MeteringPoints.Application.Create
 {
     public record ExchangeDetails(string? SourceGridAreaCode, string? TargetGridAreaCode);
 }
